+++ linter/src/rules/adding_foreign_key_constraint.rs
@@ -1,6 +1,6 @@
 use crate::violations::{RuleViolation, RuleViolationKind};
 use squawk_parser::ast::{
-    AlterTableCmds, AlterTableDef, AlterTableType, ConstrType, RootStmt, Stmt, TableElt,
+    AlterTableCmds, AlterTableDef, AlterTableType, ConstrType, RawStmt, Stmt, TableElt,
 };
 
 /// Adding a foreign key constraint requires a table scan and a
@@ -10,9 +10,9 @@ use squawk_parser::ast::{
 /// VALIDATE in another transaction will allow writes when adding the
 /// constraint.
 #[must_use]
-pub fn adding_foreign_key_constraint(tree: &[RootStmt]) -> Vec<RuleViolation> {
+pub fn adding_foreign_key_constraint(tree: &[RawStmt]) -> Vec<RuleViolation> {
     let mut errs = vec![];
-    for RootStmt::RawStmt(raw_stmt) in tree {
+    for raw_stmt in tree {
         match &raw_stmt.stmt {
             Stmt::CreateStmt(stmt) => {
                 for elt in &stmt.table_elts {