+++ linter/src/rules/adding_not_null_field.rs
@@ -1,7 +1,7 @@
 use crate::violations::{RuleViolation, RuleViolationKind};
 use crate::ViolationMessage;
 use squawk_parser::ast::{
-    AlterTableCmds, AlterTableDef, AlterTableType, ColumnDefConstraint, ConstrType, RootStmt, Stmt,
+    AlterTableCmds, AlterTableDef, AlterTableType, ColumnDefConstraint, ConstrType, RawStmt, Stmt,
 };
 
 fn has_null_and_no_default_constraint(constraints: &[ColumnDefConstraint]) -> bool {
@@ -19,9 +19,9 @@ fn has_null_and_no_default_constraint(constraints: &[ColumnDefConstraint]) -> bo
 }
 
 #[must_use]
-pub fn adding_not_nullable_field(tree: &[RootStmt]) -> Vec<RuleViolation> {
+pub fn adding_not_nullable_field(tree: &[RawStmt]) -> Vec<RuleViolation> {
     let mut errs = vec![];
-    for RootStmt::RawStmt(raw_stmt) in tree {
+    for raw_stmt in tree {
         match &raw_stmt.stmt {
             Stmt::AlterTableStmt(stmt) => {
                 for AlterTableCmds::AlterTableCmd(cmd) in &stmt.cmds {