+++ linter/src/rules/disallow_unique_constraint.rs
@@ -1,17 +1,17 @@
 use crate::rules::utils::tables_created_in_transaction;
 use crate::violations::{RuleViolation, RuleViolationKind};
 use squawk_parser::ast::{
-    AlterTableCmds, AlterTableDef, AlterTableType, ConstrType, RelationKind, RootStmt, Stmt,
+    AlterTableCmds, AlterTableDef, AlterTableType, ConstrType, RawStmt, Stmt,
 };
 
 #[must_use]
-pub fn disallow_unique_constraint(tree: &[RootStmt]) -> Vec<RuleViolation> {
+pub fn disallow_unique_constraint(tree: &[RawStmt]) -> Vec<RuleViolation> {
     let tables_created = tables_created_in_transaction(tree);
     let mut errs = vec![];
-    for RootStmt::RawStmt(raw_stmt) in tree {
+    for raw_stmt in tree {
         match &raw_stmt.stmt {
             Stmt::AlterTableStmt(stmt) => {
-                let RelationKind::RangeVar(range) = &stmt.relation;
+                let range = &stmt.relation;
                 let tbl_name = &range.relname;
                 for AlterTableCmds::AlterTableCmd(cmd) in &stmt.cmds {
                     match (&cmd.def, &cmd.subtype) {