+++ linter/src/rules/renaming_table.rs
@@ -1,10 +1,10 @@
 use crate::violations::{RuleViolation, RuleViolationKind};
-use squawk_parser::ast::{ObjectType, RootStmt, Stmt};
+use squawk_parser::ast::{ObjectType, RawStmt, Stmt};
 
 #[must_use]
-pub fn renaming_table(tree: &[RootStmt]) -> Vec<RuleViolation> {
+pub fn renaming_table(tree: &[RawStmt]) -> Vec<RuleViolation> {
     let mut errs = vec![];
-    for RootStmt::RawStmt(raw_stmt) in tree {
+    for raw_stmt in tree {
         match &raw_stmt.stmt {
             Stmt::RenameStmt(stmt) => match stmt.rename_type {
                 ObjectType::Table => {