+++ linter/src/rules/require_concurrent_index_creation.rs
@@ -1,15 +1,15 @@
 use crate::rules::utils::tables_created_in_transaction;
 use crate::violations::{RuleViolation, RuleViolationKind};
-use squawk_parser::ast::{ObjectType, RelationKind, RootStmt, Stmt};
+use squawk_parser::ast::{ObjectType, RawStmt, Stmt};
 
 #[must_use]
-pub fn require_concurrent_index_creation(tree: &[RootStmt]) -> Vec<RuleViolation> {
+pub fn require_concurrent_index_creation(tree: &[RawStmt]) -> Vec<RuleViolation> {
     let tables_created = tables_created_in_transaction(tree);
     let mut errs = vec![];
-    for RootStmt::RawStmt(raw_stmt) in tree {
+    for raw_stmt in tree {
         match &raw_stmt.stmt {
             Stmt::IndexStmt(stmt) => {
-                let RelationKind::RangeVar(range) = &stmt.relation;
+                let range = &stmt.relation;
                 let tbl_name = &range.relname;
                 if !stmt.concurrent && !tables_created.contains(tbl_name) {
                     errs.push(RuleViolation::new(