+++ parser/src/parse.rs
@@ -1,11 +1,11 @@
-use crate::ast::RootStmt;
+use crate::ast::{RawStmt, StmtRoot};
 use crate::error::PgQueryError;
 use libpg_query::{pg_query_free_parse_result, pg_query_parse};
 use serde::Deserialize;
 use serde_json::Value;
 use std::ffi::{CStr, CString};
 
-fn parse_sql_query_base<'a, T>(query: &'a str) -> Result<Vec<T>, PgQueryError>
+fn parse_sql_query_base<'a, T>(query: &'a str) -> Result<T, PgQueryError>
 where
     T: Deserialize<'a>,
 {
@@ -32,12 +32,13 @@ where
     output
 }
 
-pub fn parse_sql_query_json(query: &str) -> Result<Vec<Value>, PgQueryError> {
+pub fn parse_sql_query_json(query: &str) -> Result<Value, PgQueryError> {
     parse_sql_query_base(query)
 }
 
-pub fn parse_sql_query(query: &str) -> Result<Vec<RootStmt>, PgQueryError> {
-    parse_sql_query_base(query)
+pub fn parse_sql_query(query: &str) -> Result<Vec<RawStmt>, PgQueryError> {
+    let parsed: StmtRoot = parse_sql_query_base(query)?;
+    Ok(parsed.stmts)
 }
 
 #[cfg(test)]
@@ -45,6 +46,13 @@ mod tests {
     use super::*;
 
     use insta::assert_debug_snapshot;
+
+    #[test]
+    fn test_parse_sql_query_json() {
+        let sql = r#"ALTER TABLE table_c ADD column c boolean GENERATED ALWAYS AS (p IS NOT NULL) STORED NOT NULL;"#;
+        let res = parse_sql_query_json(sql);
+        assert_debug_snapshot!(res);
+    }
 
     #[test]
     fn test_span_with_indent() {
@@ -1238,6 +1246,16 @@ ALTER TABLE public.tasks VALIDATE CONSTRAINT tasks_fk;
 ALTER TABLE "table" ALTER CONSTRAINT "constraint" DEFERRABLE INITIALLY IMMEDIATE;
 
 ALTER TABLE "table" ALTER CONSTRAINT "constraint" NOT DEFERRABLE;
+"#;
+        let res = parse_sql_query(sql);
+        assert!(res.is_ok());
+        assert_debug_snapshot!(res);
+    }
+
+    #[test]
+    fn parse_generated_column() {
+        let sql = r#"
+ALTER TABLE table_c ADD column c boolean GENERATED ALWAYS AS (p IS NOT NULL) STORED NOT NULL;
 "#;
         let res = parse_sql_query(sql);
         assert!(res.is_ok());